+++ glados-core/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "glados-core"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+clap = "4.0.22"
+discv5 = "0.1.0"
+ethereum-types = "0.14.0"
+jsonrpc = "0.13.0"
+serde = "1.0.147"
+serde_json = "1.0.87"
+thiserror = "1.0.37"
