+++ glados-web/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod routes;
+pub mod state;
+pub mod templates;
